Added 2 files and made changes to 2 files
Added coach model file and a player model file and changed the UI of the
new coach page and the new player page. also instaled a snippet coder util
package for the dynamic forms on the new player and coach pages. Added needed
feilds to the new player and coach pages.